Fix and clarify web fallback stores after cleanup
